[FIX] https://github.com/Effective-Tilinkle/racingcar-with-java/pull/9#discussion_r1111934623
