+++ src/main/java/application/service/CarRaceService.java
@@ -2,21 +2,22 @@ package application.service;
 
 import domain.CarRace;
 import domain.CarRaceDifficulty;
-import view.View;
+import view.InputView;
+import view.OutputView;
 import view.vo.RaceCondition;
 
 public class CarRaceService {
 
 	public static void main(String[] args) {
 
-		View view = View.start();
+		InputView view = InputView.start();
 		RaceCondition raceCondition = RaceCondition.from(view);
 
 		CarRace race = CarRace.readyForRace(raceCondition, CarRaceDifficulty.NORMAL);
 
 		while (race.getNo() <= race.getTryCount()) {
 			race.race();
-			View.print(race);
+			OutputView.print(race);
 			race.countNo();
 		}
 	}
